refactor(query): split the per-key aggregate reads into their own module
Moves the shared carrier driver and all three per-key entry points out of
query.rs into `operations/get/aggregate_per_key/`, one file per concern:

  aggregate_per_key/
    mod.rs           leaf-vs-carrier vocabulary, module map
    carrier.rs       the one carrier walk every axis shares
    count.rs         query_aggregate_count_per_key
    sum.rs           query_aggregate_sum_per_key
    count_and_sum.rs query_aggregate_count_and_sum_per_key

One module per axis mirrors how the proof side already splits
`operations::proof::aggregate_{count,sum,count_and_sum}`, so the trusted
and verified surfaces are now organized the same way. Each axis file holds
only its own version gate, shape validation, leaf-shape delegation, and
error text; everything else lives in carrier.rs.

query.rs is 170 lines shorter than before this PR started — it no longer
carries any per-key aggregate code at all.

Purely a code move plus visibility/import adjustments: the driver is
`pub(super)` (visible to its sibling axis modules), and the three entry
points stay `pub` inherent methods on GroveDb, so the public API and every
call site are unchanged. `operations::get` is already minimal-gated at the
`operations` level, so the new files need no per-item cfg.

Also addresses CodeRabbit on #803: the empty-carrier tests claimed the
proved path agrees but never called prove_query. Fixed by adding the
missing differential assertion rather than deleting the claim — empty
result sets are exactly where trusted and proved could silently diverge,
and they do in fact agree.
